Move internal and file error handling into dedicated object
